Create README.md
According to the requirements of this assignment, my duty was creating SignUpForm for the user to gather information(data) from them such as username, email, password, and confirmation password by using the android application. However, to implement this, I created the TextView and EditText XML  which make the user entered their information in it easily. Then, I created the button that used to display entered information on the screen when the user pressed it. In spite, when the user entered valid information the program should print( show) them the "Welcome" message, otherwise, if the user entered invalid information like a password when the program asked them "confirmation Password", they will see the message on the screen says"Password does not match".